+++ newIDE/app/src/InstancesEditor/TileMapPaintingPreview.js
@@ -9,13 +9,12 @@ import RenderedInstance from '../ObjectsRendering/Renderers/RenderedInstance';
 import Rendered3DInstance from '../ObjectsRendering/Renderers/Rendered3DInstance';
 import { type TileMapTileSelection } from './TileSetVisualizer';
 import { AffineTransformation } from '../Utils/AffineTransformation';
-
-type TileSet = {|
-  rowCount: number,
-  columnCount: number,
-  tileSize: number,
-  atlasImage: string,
-|};
+import {
+  getTileSet,
+  getTilesGridCoordinatesFromPointerSceneCoordinates,
+  isTileSetBadlyConfigured,
+  type TileSet,
+} from '../Utils/TileMap';
 
 export const updateSceneToTileMapTransformation = (
   instance: gdInitialInstance,
@@ -67,291 +66,6 @@ export const updateSceneToTileMapTransformation = (
   sceneToTileMapTransformation.copyFrom(tileMapToSceneTransformation);
   sceneToTileMapTransformation.invert();
   return { scaleX, scaleY };
-};
-
-export const getTileSet = (object: gdObject): TileSet => {
-  const objectConfigurationProperties = object
-    .getConfiguration()
-    .getProperties();
-  const columnCount = parseFloat(
-    objectConfigurationProperties.get('columnCount').getValue()
-  );
-  const rowCount = parseFloat(
-    objectConfigurationProperties.get('rowCount').getValue()
-  );
-  const tileSize = parseFloat(
-    objectConfigurationProperties.get('tileSize').getValue()
-  );
-  const atlasImage = objectConfigurationProperties.get('atlasImage').getValue();
-  return { rowCount, columnCount, tileSize, atlasImage };
-};
-
-export const isTileSetBadlyConfigured = ({
-  rowCount,
-  columnCount,
-  tileSize,
-  atlasImage,
-}: TileSet) => {
-  return (
-    !Number.isInteger(columnCount) ||
-    columnCount <= 0 ||
-    !Number.isInteger(rowCount) ||
-    rowCount <= 0
-  );
-};
-
-export const isSelectionASingleTileRectangle = (
-  tileMapTileSelection: TileMapTileSelection
-): boolean => {
-  return (
-    tileMapTileSelection.kind === 'rectangle' &&
-    tileMapTileSelection.coordinates.length === 2 &&
-    tileMapTileSelection.coordinates[0].x ===
-      tileMapTileSelection.coordinates[1].x &&
-    tileMapTileSelection.coordinates[0].y ===
-      tileMapTileSelection.coordinates[1].y
-  );
-};
-
-/**
- * When flipping is activated, the tile texture should be flipped but
- * the tiles should be flipped as well in the selection
- * (the left tiles should be replaced by the right tiles if the horizontal flip
- * is activated).
- */
-const getTileCorrespondingToFlippingInstructions = ({
-  tileMapTileSelection,
-  tileCoordinates,
-}: {|
-  tileMapTileSelection: TileMapTileSelection,
-  tileCoordinates: {| x: number, y: number |},
-|}): {| x: number, y: number |} => {
-  if (tileMapTileSelection.kind === 'rectangle') {
-    const selectionTopLeftCorner = tileMapTileSelection.coordinates[0];
-    const selectionBottomRightCorner = tileMapTileSelection.coordinates[1];
-    const selectionWidth =
-      selectionBottomRightCorner.x - selectionTopLeftCorner.x + 1;
-    const selectionHeight =
-      selectionBottomRightCorner.y - selectionTopLeftCorner.y + 1;
-    const deltaX = tileCoordinates.x - selectionTopLeftCorner.x;
-    const deltaY = tileCoordinates.y - selectionTopLeftCorner.y;
-    const newX =
-      selectionTopLeftCorner.x +
-      (tileMapTileSelection.flipHorizontally
-        ? selectionWidth - deltaX - 1
-        : deltaX);
-    const newY =
-      selectionTopLeftCorner.y +
-      (tileMapTileSelection.flipVertically
-        ? selectionHeight - deltaY - 1
-        : deltaY);
-    return { x: newX, y: newY };
-  }
-  return tileCoordinates;
-};
-
-/**
- * Returns the list of tiles corresponding to the user selection.
- */
-export const getTilesGridCoordinatesFromPointerSceneCoordinates = ({
-  tileMapTileSelection,
-  coordinates,
-  tileSize,
-  sceneToTileMapTransformation,
-}: {|
-  tileMapTileSelection: TileMapTileSelection,
-  coordinates: Array<{| x: number, y: number |}>,
-  tileSize: number,
-  sceneToTileMapTransformation: AffineTransformation,
-|}): {|
-  tileCoordinates?: {| x: number, y: number |},
-  erase?: boolean,
-  topLeftCorner: {| x: number, y: number |},
-  bottomRightCorner: {| x: number, y: number |},
-|}[] => {
-  if (coordinates.length === 0) return [];
-
-  const tilesCoordinatesInTileMapGrid = [];
-  if (coordinates.length === 1) {
-    // One coordinate corresponds to the pointer over the canvas.
-    const coordinatesInTileMapGrid = [0, 0];
-    sceneToTileMapTransformation.transform(
-      [coordinates[0].x, coordinates[0].y],
-      coordinatesInTileMapGrid
-    );
-    const x = Math.floor(coordinatesInTileMapGrid[0] / tileSize);
-    const y = Math.floor(coordinatesInTileMapGrid[1] / tileSize);
-    let tileCoordinates;
-    if (tileMapTileSelection.kind === 'rectangle') {
-      const topLeftCorner = tileMapTileSelection.coordinates[0];
-      tileCoordinates = getTileCorrespondingToFlippingInstructions({
-        tileMapTileSelection,
-        tileCoordinates: topLeftCorner,
-      });
-    }
-    return [
-      {
-        erase: tileMapTileSelection.kind === 'erase',
-        tileCoordinates,
-        topLeftCorner: { x, y },
-        bottomRightCorner: { x, y },
-      },
-    ];
-  }
-
-  if (coordinates.length === 2) {
-    const firstPointCoordinatesInTileMap = [0, 0];
-    sceneToTileMapTransformation.transform(
-      [coordinates[0].x, coordinates[0].y],
-      firstPointCoordinatesInTileMap
-    );
-    const secondPointCoordinatesInTileMap = [0, 0];
-    sceneToTileMapTransformation.transform(
-      [coordinates[1].x, coordinates[1].y],
-      secondPointCoordinatesInTileMap
-    );
-    const topLeftCornerCoordinatesInTileMap = [
-      Math.min(
-        firstPointCoordinatesInTileMap[0],
-        secondPointCoordinatesInTileMap[0]
-      ),
-      Math.min(
-        firstPointCoordinatesInTileMap[1],
-        secondPointCoordinatesInTileMap[1]
-      ),
-    ];
-    const bottomRightCornerCoordinatesInTileMap = [
-      Math.max(
-        firstPointCoordinatesInTileMap[0],
-        secondPointCoordinatesInTileMap[0]
-      ),
-      Math.max(
-        firstPointCoordinatesInTileMap[1],
-        secondPointCoordinatesInTileMap[1]
-      ),
-    ];
-    const topLeftCornerCoordinatesInTileMapGrid = [
-      Math.floor(topLeftCornerCoordinatesInTileMap[0] / tileSize),
-      Math.floor(topLeftCornerCoordinatesInTileMap[1] / tileSize),
-    ];
-    const bottomRightCornerCoordinatesInTileMapGrid = [
-      Math.floor(bottomRightCornerCoordinatesInTileMap[0] / tileSize),
-      Math.floor(bottomRightCornerCoordinatesInTileMap[1] / tileSize),
-    ];
-    if (tileMapTileSelection.kind === 'erase') {
-      tilesCoordinatesInTileMapGrid.push({
-        erase: true,
-        topLeftCorner: {
-          x: topLeftCornerCoordinatesInTileMapGrid[0],
-          y: topLeftCornerCoordinatesInTileMapGrid[1],
-        },
-        bottomRightCorner: {
-          x: bottomRightCornerCoordinatesInTileMapGrid[0],
-          y: bottomRightCornerCoordinatesInTileMapGrid[1],
-        },
-      });
-    } else if (tileMapTileSelection.kind === 'rectangle') {
-      const selectionTopLeftCorner = tileMapTileSelection.coordinates[0];
-      const selectionBottomRightCorner = tileMapTileSelection.coordinates[1];
-      const selectionWidth =
-        selectionBottomRightCorner.x - selectionTopLeftCorner.x + 1;
-      const selectionHeight =
-        selectionBottomRightCorner.y - selectionTopLeftCorner.y + 1;
-
-      if (isSelectionASingleTileRectangle(tileMapTileSelection)) {
-        tilesCoordinatesInTileMapGrid.push({
-          tileCoordinates: getTileCorrespondingToFlippingInstructions({
-            tileMapTileSelection,
-            tileCoordinates: selectionTopLeftCorner,
-          }),
-          topLeftCorner: {
-            x: topLeftCornerCoordinatesInTileMapGrid[0],
-            y: topLeftCornerCoordinatesInTileMapGrid[1],
-          },
-          bottomRightCorner: {
-            x: bottomRightCornerCoordinatesInTileMapGrid[0],
-            y: bottomRightCornerCoordinatesInTileMapGrid[1],
-          },
-        });
-        return tilesCoordinatesInTileMapGrid;
-      }
-
-      for (
-        let x = topLeftCornerCoordinatesInTileMapGrid[0];
-        x <= bottomRightCornerCoordinatesInTileMapGrid[0];
-        x++
-      ) {
-        for (
-          let y = topLeftCornerCoordinatesInTileMapGrid[1];
-          y <= bottomRightCornerCoordinatesInTileMapGrid[1];
-          y++
-        ) {
-          const deltaX = x - topLeftCornerCoordinatesInTileMapGrid[0];
-          const deltaY = y - topLeftCornerCoordinatesInTileMapGrid[1];
-          const invertedDeltaX =
-            bottomRightCornerCoordinatesInTileMapGrid[0] - x;
-          const invertedDeltaY =
-            bottomRightCornerCoordinatesInTileMapGrid[1] - y;
-          if (deltaX === 0 && deltaY === 0) {
-            tilesCoordinatesInTileMapGrid.push({
-              tileCoordinates: getTileCorrespondingToFlippingInstructions({
-                tileMapTileSelection,
-                tileCoordinates: selectionTopLeftCorner,
-              }),
-              topLeftCorner: { x, y },
-              bottomRightCorner: { x, y },
-            });
-            continue;
-          }
-          if (invertedDeltaX === 0 && invertedDeltaY === 0) {
-            tilesCoordinatesInTileMapGrid.push({
-              tileCoordinates: getTileCorrespondingToFlippingInstructions({
-                tileMapTileSelection,
-                tileCoordinates: selectionBottomRightCorner,
-              }),
-              topLeftCorner: { x, y },
-              bottomRightCorner: { x, y },
-            });
-            continue;
-          }
-
-          let tileX, tileY;
-          if (deltaX === 0 || selectionWidth === 1) {
-            tileX = selectionTopLeftCorner.x;
-          } else if (invertedDeltaX === 0 || selectionWidth === 2) {
-            tileX = selectionBottomRightCorner.x;
-          } else {
-            tileX =
-              ((deltaX - 1) % (selectionWidth - 2)) +
-              1 +
-              selectionTopLeftCorner.x;
-          }
-          if (deltaY === 0 || selectionHeight === 1) {
-            tileY = selectionTopLeftCorner.y;
-          } else if (invertedDeltaY === 0 || selectionHeight === 2) {
-            tileY = selectionBottomRightCorner.y;
-          } else {
-            tileY =
-              ((deltaY - 1) % (selectionHeight - 2)) +
-              1 +
-              selectionTopLeftCorner.y;
-          }
-
-          tilesCoordinatesInTileMapGrid.push({
-            tileCoordinates: getTileCorrespondingToFlippingInstructions({
-              tileMapTileSelection,
-              tileCoordinates: { x: tileX, y: tileY },
-            }),
-            topLeftCorner: { x, y },
-            bottomRightCorner: { x, y },
-          });
-        }
-      }
-    }
-  }
-  // TODO: Optimize tiles by stretching the tiling textures instead of creating new ones
-  // for each coordinate.
-  return tilesCoordinatesInTileMapGrid;
 };
 
 type Props = {|
